Migrate hook callback methods to native.Manticore
