+++ manticore/native/manticore.py
@@ -1,4 +1,5 @@
 import logging
+import types
 
 import elftools
 import os
@@ -73,6 +74,64 @@ class Manticore(ManticoreBase):
             return cls(_make_decree(path, concrete_start), **kwargs)
         except KeyError:  # FIXME(mark) magic parsing for DECREE should raise better error
             raise Exception(f'Invalid binary: {path}')
+
+    def init(self, f):
+        '''
+        A decorator used to register a hook function to run before analysis begins. Hook
+        function takes one :class:`~manticore.core.state.State` argument.
+        '''
+        def callback(manticore_obj, state):
+                       f(state)
+        self.subscribe('will_start_run', types.MethodType(callback, self))
+        return f
+
+    def hook(self, pc):
+        '''
+        A decorator used to register a hook function for a given instruction address.
+        Equivalent to calling :func:`~add_hook`.
+
+        :param pc: Address of instruction to hook
+        :type pc: int or None
+        '''
+        def decorator(f):
+            self.add_hook(pc, f)
+            return f
+        return decorator
+
+    def add_hook(self, pc, callback):
+        '''
+        Add a callback to be invoked on executing a program counter. Pass `None`
+        for pc to invoke callback on every instruction. `callback` should be a callable
+        that takes one :class:`~manticore.core.state.State` argument.
+
+        :param pc: Address of instruction to hook
+        :type pc: int or None
+        :param callable callback: Hook function
+        '''
+        if not (isinstance(pc, int) or pc is None):
+            raise TypeError(f"pc must be either an int or None, not {pc.__class__.__name__}")
+        else:
+            self._hooks.setdefault(pc, set()).add(callback)
+            if self._hooks:
+                self._executor.subscribe('will_execute_instruction', self._hook_callback)
+
+    def _hook_callback(self, state, pc, instruction):
+        'Invoke all registered generic hooks'
+
+        # Ignore symbolic pc.
+        # TODO(yan): Should we ask the solver if any of the hooks are possible,
+        # and execute those that are?
+
+        if issymbolic(pc):
+                        return
+
+        # Invoke all pc-specific hooks
+        for cb in self._hooks.get(pc, []):
+                        cb(state)
+
+        # Invoke all pc-agnostic hooks
+        for cb in self._hooks.get(None, []):
+                        cb(state)
 
     #############################################################################
     #############################################################################
